requirements.txt
autopep8==1.5
click==7.1.1
colorama==0.4.3
django-qrcode==0.3
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycodestyle==2.5.0
python-dotenv==0.12.0
pytz==2019.3
pyzbar==0.1.8
qrcode==6.1
qrtools==0.0.2
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0